ci(general): use binary releases over Docker images for Taplo
It seems that Taplo changed its Docker image from an Alpine image to a
just a binary with version 0.10.0. The `toml` GitHub Actions job
previously ran in the Alpine environment of the image, so this broke
that job. This commit changes the job to download Taplo's binary
directly from its GitHub releases using `taiki-e/install-action@v2`.
This will be followed by changes to how it's checked for updates. The
same change will probably be made for `shfmt` soon as well.