ci(updates): generalize GitHub Actions version checking to any releases
Generalized the version checking for GitHub Actions releases to check
for any GitHub repository's releases. There wasn't actually any real
work needed to do this, the mechanism _should_ be the same, but adding
Taplo into the list of repositories to check revealed that comes from
weird tag names. This was fixed by filter tag names better. This commit
also changes the API endpoint to one specific to releases instead of any
tag, because Git tags don't have to be associated with releases (and the
releases endpoint is more useful because it puts the latest first).