refactor: generate a Markdown file with dependency licenses
Uses [`license_page`](https://github.com/RemasteredArch/license_page)
(currently at `c19f73e`) in a build script to generate a Markdown file
listing the licenses (and the full text of said licenses) of all of
1N4's dependencies. This also adds a couple exemptions to `deny.toml`,
one of which will need to be removed when a PR lands in upstream
`cargo-license`. This commit doesn't really change any code, but I
marked it as a `refactor` because it's not being used in anything yet.