fix(build): don't include licenses of `dev-dependencies` in `licenses.md`
Previously, `license_page::opt::GetLicensesOpt::avoid_dev_deps` would
be set so long as `PROFILE != "debug"`, based on the assumption that
"dev deps" referred to dependencies referred to debug-mode only
dependencies (because "debug mode" is another name for the "dev profile"
in Cargo), but it actually refers to licenses only used for "tests,
examples, and benchmarks," so it is now set to always avoid them. See
new comment for more details.