feat!: add options to disable file and console logging individually
Added `--disable-file-logging` and `--disable-console-logging` to add
fine-grained control over logger output. This also comes along with
moving `--quiet` out of `ina_logging::setting::Settings`, a breaking
change. `--disable-file-logging` is set by default in the Docker
container because Docker has its own logging facilities, and writing to
files that are shared with the host is annoying for non-root container
users.